refactor: comment entity 배럴 파일 적용 및 절대 경로 alias로 import 변경
