+++ src/entities/comment/createComment/createComment.api.ts
@@ -1,7 +1,7 @@
 import { http } from "../../../shared/api"
-import { CreateComment } from "./createComment.model"
+import { CreateCommentRequest } from "./createComment.model"
 import { CommentResponse } from "../../comment/model/commentResponse"
 
-export const createComment = async (comment: CreateComment): Promise<CommentResponse> => {
+export const createComment = async (comment: CreateCommentRequest): Promise<CommentResponse> => {
   return await http.post("/api/comments/add", { ...comment })
 }
