+++ src/entities/comment/createComment/createComment.model.ts
@@ -1,4 +1,4 @@
-export interface CreateComment {
+export interface CreateCommentRequest {
   postId: number
   body: string
   userId: number
