+++ src/entities/comment/createComment/index.ts
@@ -1,0 +1,2 @@
+export * from "./createComment.api"
+export * from "./createComment.model"
