+++ src/entities/comment/deleteComment/index.ts
@@ -1,0 +1,2 @@
+export * from "./deleteComment.api"
+export * from "./deleteComment.model"
