+++ src/entities/comment/getComment/getComment.api.ts
@@ -1,6 +1,6 @@
 import { http } from "../../../shared/api"
-import { CommentList } from "./getComment.model"
+import { CommentListResponse } from "./getComment.model"
 
-export const getComments = async (postId: number): Promise<CommentList> => {
+export const getComments = async (postId: number): Promise<CommentListResponse> => {
   return await http.get(`/api/comments/post/${postId}`)
 }
