+++ src/entities/comment/getComment/index.ts
@@ -1,0 +1,2 @@
+export * from "./getComment.api"
+export * from "./getComment.model"
