+++ src/entities/comment/likeComment/index.ts
@@ -1,0 +1,2 @@
+export * from "./likeComment.api"
+export * from "./likeComment.model"
