+++ src/entities/comment/updateComment/index.ts
@@ -1,0 +1,1 @@
+export * from "./updateComment.api"
