+++ src/entities/comment/updateComment/updateComment.model.ts
@@ -1,0 +1,4 @@
+export interface UpdateCommentRequest {
+  id: number
+  body: string
+}
