--- src/features/comment/createComment/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateCommentButton"
