+++ src/features/comment/deleteComment/api/useDeleteCommentQuery.ts
@@ -1,8 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { DeleteCommentResponse } from "../../../entities/comment/deleteComment/deleteComment.model"
-import { CommentList } from "../../../entities/comment/getComment/getComment.model"
-import { deleteComment } from "../../../entities/comment/deleteComment/deleteComment.api"
-import { QUERY_KEYS } from "../config/queryKeys"
+import { DeleteCommentResponse, CommentListResponse, deleteComment } from "@entities/comment"
+import { QUERY_KEYS } from "@features/comment/config/queryKeys"
 
 export const useDeleteCommentQuery = () => {
   const queryClient = useQueryClient()
@@ -13,7 +11,7 @@ export const useDeleteCommentQuery = () => {
       // postId를 키에 포함해서 해당 댓글 리스트만 갱신
       const key = QUERY_KEYS.GET_COMMENT(deleted.postId)
 
-      queryClient.setQueryData<CommentList>(key, (old) => {
+      queryClient.setQueryData<CommentListResponse>(key, (old) => {
         if (!old) {
           // 캐시가 비어 있으면 기본값 반환
           return {
