+++ src/features/comment/getComment/api/useGetCommentsQuery.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from "@tanstack/react-query"
-import { getComments } from "../../../entities/comment/getComment/getComment.api"
-import { QUERY_KEYS } from "../config/queryKeys"
+import { getComments } from "@entities/comment"
+import { QUERY_KEYS } from "@features/comment/config/queryKeys"
 
 export const useGetCommentsQuery = (postId: number) => {
   const { data: comments } = useSuspenseQuery({
