+++ src/features/comment/likeComment/api/useLikeCommentQuery.ts
@@ -1,8 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { likeComment } from "../../../entities/comment/likeComment/likeComment.api"
-import { LikeCommentResponse } from "../../../entities/comment/likeComment/likeComment.model"
-import { CommentList } from "../../../entities/comment/getComment/getComment.model"
-import { QUERY_KEYS } from "../config/queryKeys"
+import { likeComment, LikeCommentResponse, CommentListResponse } from "@entities/comment"
+import { QUERY_KEYS } from "@features/comment/config/queryKeys"
 
 export const useLikeCommentQuery = () => {
   const queryClient = useQueryClient()
@@ -12,7 +10,7 @@ export const useLikeCommentQuery = () => {
     onSuccess: (liked: LikeCommentResponse) => {
       const key = QUERY_KEYS.GET_COMMENT(liked.postId)
 
-      queryClient.setQueryData<CommentList>(key, (old) => {
+      queryClient.setQueryData<CommentListResponse>(key, (old) => {
         if (!old) {
           return {
             comments: [],
