+++ src/features/comment/updateComment/ui/UpdateCommentForm.tsx
@@ -1,6 +1,5 @@
-import { Button, Textarea } from "../../../../shared/ui"
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../../shared/ui/dialog"
-import { useUpdateCommentForm } from "../updateComment.model"
+import { Button, Textarea, Dialog, DialogContent, DialogHeader, DialogTitle } from "@shared/ui"
+import { useUpdateCommentForm } from "@features/comment/updateComment/model/useUpdateCommentForm"
 
 interface UpdateCommentFormProps {
   isOpen: boolean
