+++ src/features/post/getPost/ui/PostDetail.tsx
@@ -1,6 +1,6 @@
 import { Post } from "../../../../entities/post/model"
 import { Dialog, DialogContent, DialogHeader, DialogTitle, HighlightText } from "../../../../shared/ui"
-import { CommentList } from "../../../comment/getComment/getComment.ui"
+import { CommentList } from "../../../comment/getComment/ui/CommentList"
 interface PostDetailProps {
   isOpen: boolean
   close: () => void
