+++ vite.config.ts
@@ -1,9 +1,19 @@
 import { defineConfig } from "vite"
 import react from "@vitejs/plugin-react"
+import path from "node:path"
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      // tsconfig 의 baseUrl: 'src' 와 매치
+      "@shared": path.resolve(__dirname, "./src/shared"),
+      "@entities": path.resolve(__dirname, "./src/entities"),
+      "@features": path.resolve(__dirname, "./src/features"),
+      "@widgets": path.resolve(__dirname, "./src/widgets"),
+    },
+  },
   server: {
     proxy: {
       "/api": {
